Finished exercise

+++ week2/model.php
@@ -97,7 +97,7 @@ function get_breadcrumbs($breadcrumbs) {
  * @param array $navigation Array with as Key the page name and as Value the corresponding url
  * @return string html code that represents the navigation
  */
-function get_navigation($navigation){
+function get_navigation($template, $active_id){
     $navigation_exp = '
     <nav class="navbar navbar-expand-lg navbar-light bg-light">
     <a class="navbar-brand">Series Overview</a>
@@ -106,13 +106,13 @@ function get_navigation($navigation){
     </button>
     <div class="collapse navbar-collapse" id="navbarSupportedContent">
     <ul class="navbar-nav mr-auto">';
-    foreach ($navigation as $name => $info) {
-        if ($info[1]){
+    foreach ($template as $key => $values) {
+        if ($key == $active_id){
             $navigation_exp .= '<li class="nav-item active">';
-            $navigation_exp .= '<a class="nav-link" href="'.$info[0].'">'.$name.'</a>';
+            $navigation_exp .= '<a class="nav-link" href="'.$values['url'].'">'.$values['name'].'</a>';
         }else{
             $navigation_exp .= '<li class="nav-item">';
-            $navigation_exp .= '<a class="nav-link" href="'.$info[0].'">'.$name.'</a>';
+            $navigation_exp .= '<a class="nav-link" href="'.$values['url'].'">'.$values['name'].'</a>';
         }
 
         $navigation_exp .= '</li>';
@@ -135,15 +135,17 @@ function get_serie_table($pdo, $series){
     <thead
     <tr>
         <th scope="col">Series</th>
+        <th scope="col">Added by</th>
         <th scope="col"></th>
     </tr>
     </thead>
     <tbody>';
     foreach($series as $key => $value){
+        $user = get_username($pdo, $value['user_id']);
         $table_exp .= '
         <tr>
             <th scope="row">'.$value['name'].'</th>
-            <th scope="row">'.get_user_name($pdo, $value['user_id']).'</th>
+            <th scope="row">'. $user['username'] .'</th>
             <td><a href="/DDWT19/week2/serie/?serie_id='.$value['id'].'" role="button" class="btn btn-primary">More info</a></td>
         </tr>
         ';
@@ -210,6 +212,7 @@ function get_serieinfo($pdo, $serie_id){
  * @return string
  */
 function get_error($feedback){
+    $feedback = json_decode($feedback, True);
     $error_exp = '
         <div class="alert alert-'.$feedback['type'].'" role="alert">
             '.$feedback['message'].'
@@ -223,7 +226,7 @@ function get_error($feedback){
  * @param array $serie_info post array
  * @return array with message feedback
  */
-function add_serie($pdo, $serie_info){
+function add_serie($pdo, $serie_info, $user_id){
     /* Check if all fields are set */
     if (
         empty($serie_info['Name']) or
@@ -257,12 +260,13 @@ function add_serie($pdo, $serie_info){
     }
 
     /* Add Serie */
-    $stmt = $pdo->prepare("INSERT INTO series (name, creator, seasons, abstract,user_id) VALUES (?, ?, ?, ?,?)");
+    $stmt = $pdo->prepare("INSERT INTO series (name, creator, seasons, abstract,user_id) VALUES (?, ?, ?, ?, ?)");
     $stmt->execute([
         $serie_info['Name'],
         $serie_info['Creator'],
         $serie_info['Seasons'],
         $serie_info['Abstract'],
+        $user_id
     ]);
     $inserted = $stmt->rowCount();
     if ($inserted ==  1) {
@@ -285,7 +289,7 @@ function add_serie($pdo, $serie_info){
  * @param array $serie_info post array
  * @return array
  */
-function update_serie($pdo, $serie_info){
+function update_serie($pdo, $serie_info, $user_id){
     /* Check if all fields are set */
     if (
         empty($serie_info['Name']) or
@@ -299,7 +303,12 @@ function update_serie($pdo, $serie_info){
             'message' => 'There was an error. Not all fields were filled in.'
         ];
     }
-
+    if ($serie_info['user_id'] != $user_id) {
+        return [
+            'type' => 'danger',
+            'message' => 'You aren\'t allow to change this'
+        ];
+    }
     /* Check data type */
     if (!is_numeric($serie_info['Seasons'])) {
         return [
@@ -326,13 +335,14 @@ function update_serie($pdo, $serie_info){
     }
 
     /* Update Serie */
-    $stmt = $pdo->prepare("UPDATE series SET name = ?, creator = ?, seasons = ?, abstract = ? WHERE id = ?");
+    $stmt = $pdo->prepare("UPDATE series SET name = ?, creator = ?, seasons = ?, abstract = ?, user_id = ? WHERE id = ?");
     $stmt->execute([
         $serie_info['Name'],
         $serie_info['Creator'],
         $serie_info['Seasons'],
         $serie_info['Abstract'],
-        $serie_info['serie_id']
+        $serie_info['serie_id'],
+        $user_id
     ]);
     $updated = $stmt->rowCount();
     if ($updated ==  1) {
@@ -355,10 +365,15 @@ function update_serie($pdo, $serie_info){
  * @param int $serie_id id of the to be deleted series
  * @return array
  */
-function remove_serie($pdo, $serie_id){
+function remove_serie($pdo, $serie_id, $user_id){
     /* Get series info */
     $serie_info = get_serieinfo($pdo, $serie_id);
-
+    if ($serie_info['user_id'] != $user_id) {
+        return [
+            'type' => 'danger',
+            'message' => 'You aren\'t allow to change this'
+        ];
+    }
     /* Delete Serie */
     $stmt = $pdo->prepare("DELETE FROM series WHERE id = ?");
     $stmt->execute([$serie_id]);
@@ -390,6 +405,14 @@ function count_series($pdo){
     return $series;
 }
 
+function count_users($pdo){
+    /* Get users */
+    $stmt = $pdo->prepare('SELECT * FROM users');
+    $stmt->execute();
+    $users = $stmt->rowCount();
+    return $users;
+}
+
 /**
  * Changes the HTTP Header to a given location
  * @param string $location location to be redirected to
@@ -413,10 +436,173 @@ function get_user_id(){
     }
 }
 
-function get_user_name($pdo, $user_id){
-    $stmt = $pdo->prepare('SELECT firstname, lastname FROM users WHERE id = ?');
+/**
+ * Get current user name
+ *
+ */
+function get_username($pdo, $user_id){
+    $stmt = $pdo->prepare('SELECT username, firstname, lastname FROM users WHERE id = ?');
     $stmt->execute([$user_id]);
-    $name = $stmt->fetch_assoc();
+    $name = $stmt->fetch();
+
     return $name;
 
+}
+
+/* check if logged in */
+function check_login(){
+    session_start();
+    if (isset($_SESSION['user_id'])){
+        return True;
+    } else {
+        return False;
+    }
+}
+
+/**
+ * Registers the user and logs the user in when successful
+ * @param object $pdo database object
+ * @param array $form_data with POST data
+ * @return array
+ */
+function register_user($pdo, $form_data)
+{
+    /* Check if all fields are set */
+    if (
+        empty($form_data['username']) or
+        empty($form_data['password']) or
+        empty($form_data['firstname']) or
+        empty($form_data['lastname'])
+    ) {
+        return [
+            'type' => 'danger',
+            'message' => 'You should enter a username, password, first- and last name.'
+        ];
+    }
+
+    /* Check if user already exists */
+    try {
+        $stmt = $pdo->prepare('SELECT * FROM users WHERE username = ?');
+        $stmt->execute([$form_data['username']]);
+        $user_exists = $stmt->rowCount();
+    } catch (\PDOException $e) {
+        return [
+            'type' => 'danger',
+            'message' => sprintf('There was an error: %s', $e->getMessage())
+        ];
+    }
+    /* Return error message for existing username */
+    if (!empty($user_exists)) {
+        return [
+            'type' => 'danger',
+            'message' => 'The username you entered does already exists!'
+        ];
+    }
+
+    /* Hash password */
+    $password = password_hash($form_data['password'], PASSWORD_DEFAULT);
+    /* Save user to the database */
+    try {
+        $stmt = $pdo->prepare('INSERT INTO users (username, password, firstname,
+            lastname) VALUES (?, ?, ?, ?)');
+        $stmt->execute([$form_data['username'], $password, $form_data['firstname'],
+            $form_data['lastname']]);
+        $user_id = $pdo->lastInsertId();
+    } catch (PDOException $e) {
+        return [
+            'type' => 'danger',
+            'message' => sprintf('There was an error: %s', $e->getMessage())
+        ];
+    }
+
+    /* Login user and redirect */
+    session_start();
+    $_SESSION['user_id'] = $user_id;
+    $feedback = [
+        'type' => 'success',
+        'message' => sprintf('%s, your account was successfully
+        created!', get_username($pdo, $_SESSION['user_id']))
+    ];
+    redirect(sprintf('/DDWT19/week2/myaccount/?error_msg=%s',
+        json_encode($feedback)));
+}
+
+/**
+ * Logs the user in when successful
+ * @param object $pdo database object
+ * @param array $form_data with POST data
+ * @return array
+ */
+
+function login_user($pdo, $form_data)
+{
+    /* Check if all fields are set */
+    if (
+        empty($form_data['username']) or
+        empty($form_data['password'])
+    ) {
+        return [
+            'type' => 'danger',
+            'message' => 'You should enter a username and password.'
+        ];
+    }
+    /* Check if user exists */
+    try {
+        $stmt = $pdo->prepare('SELECT * FROM users WHERE username = ?');
+        $stmt->execute([$form_data['username']]);
+        $user_info = $stmt->fetch();
+    } catch (\PDOException $e) {
+        return [
+            'type' => 'danger',
+            'message' => sprintf('There was an error: %s', $e->getMessage())
+        ];
+    }
+    /* Return error message for wrong username */
+    if (empty($user_info)) {
+        return [
+            'type' => 'danger',
+            'message' => 'The username you entered does not exist!'
+        ];
+    }
+    /* Check password */
+    if (!password_verify($form_data['password'], $user_info['password'])) {
+        return [
+            'type' => 'danger',
+            'message' => 'The password you entered is incorrect!'
+        ];
+    } else {
+        session_start();
+        $_SESSION['user_id'] = $user_info['id'];
+        $feedback = [
+            'type' => 'success',
+            'message' => sprintf('%s, you were logged in successfully!',
+                get_username($pdo, $_SESSION['user_id'])['firstname'])
+        ];
+        redirect(sprintf('/DDWT19/week2/myaccount/?error_msg=%s',
+            json_encode($feedback)));
+    }
+}
+
+/**
+ * Logs the user out when successful
+ * @return array
+ */
+
+function logout_user(){
+    session_start();
+    if ( empty($_SESSION['user_id']))
+    {
+        return [
+            'type' => 'danger',
+            'message' => 'You weren\'t even logged in!'
+        ];
+    }
+    else
+    {
+        session_destroy();
+        return [
+            'type' => 'success',
+            'message' => 'You were successfully signed out!'
+        ];
+    }
 }
